#13-Exception 케이스 추가

+++ src/main/java/com/fofo/core/domain/match/MatchRemover.java
@@ -1,6 +1,8 @@
 package com.fofo.core.domain.match;
 
 import com.fofo.core.storage.MatchRepository;
+import com.fofo.core.support.error.CoreApiException;
+import com.fofo.core.support.error.CoreErrorType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -13,6 +15,9 @@ public class MatchRemover {
     private final MatchRepository matchRepository;
 
     public void removeMatch(final List<Long> matchIdList) {
+        if(!matchRepository.findUnCancelableMatches(matchIdList).isEmpty()){
+            throw new CoreApiException(CoreErrorType.MATCH_UNCANCELABLE_ERROR);
+        };
         matchRepository.deleteMatchesBy(matchIdList);
     }
 }
